+++ bloxlink_lib/__init__.py
@@ -1,7 +1,6 @@
 import logging
 
 from .models.base import *
-from .models.schemas import *
 from .models.roblox import *
 from .models.binds import *
 from .exceptions import *
