+++ bloxlink_lib/database/redis.py
@@ -10,7 +10,7 @@ from redis.asyncio import Redis
 from redis import ConnectionError as RedisConnectionError
 
 from bloxlink_lib import BaseModel
-from .config import CONFIG
+from bloxlink_lib.config import CONFIG
 
 redis: Redis = None
 
@@ -42,15 +42,25 @@ def connect_redis():
     # loop.create_task(_heartbeat_loop()) # TODO: fix this
 
 
-async def redis_set(key: str, value: BaseModel | Any, expire: datetime.timedelta | int = None, **kwargs):
+async def redis_set(
+    key: str, value: BaseModel | Any, expire: datetime.timedelta | int = None, **kwargs
+):
     """Set a value in Redis. Accepts BaseModels and expirations as datetimes."""
 
-    await redis._old_set(key,  # pylint: disable=protected-access
-                         value.model_dump_json() if isinstance(value, BaseModel) else (
-                             json.dumps(value) if isinstance(value, (list, dict)) else value),
-                         ex=int(expire.total_seconds()) if expire and isinstance(
-                             expire, datetime.timedelta) else expire,
-                         **kwargs)
+    await redis._old_set(
+        key,  # pylint: disable=protected-access
+        (
+            value.model_dump_json()
+            if isinstance(value, BaseModel)
+            else (json.dumps(value) if isinstance(value, (list, dict)) else value)
+        ),
+        ex=(
+            int(expire.total_seconds())
+            if expire and isinstance(expire, datetime.timedelta)
+            else expire
+        ),
+        **kwargs,
+    )
 
 
 async def _heartbeat_loop():
