+++ bloxlink_lib/models/migrators.py
@@ -4,8 +4,8 @@ from bloxlink_lib.models.schemas.guilds.guild_types import GuildRestriction
 
 
 def migrate_restrictions(
-    restrictions: dict[str, dict[str, guilds.GuildRestriction]]
-) -> list[guilds.GuildRestriction]:
+    restrictions: dict[str, dict[str, GuildRestriction]]
+) -> list[GuildRestriction]:
     if isinstance(restrictions, list):
         return restrictions
 
@@ -14,7 +14,7 @@ def migrate_restrictions(
     for restriction_type, restriction_data in restrictions.items():
         for restricted_id, restriction in restriction_data.items():
             new_guild_restrictions.append(
-                guilds.GuildRestriction(
+                GuildRestriction(
                     id=restricted_id,
                     displayName=restriction["name"],
                     addedBy=restriction["addedBy"],
