+++ bloxlink_lib/models/schemas/__init__.py
@@ -1,7 +1,7 @@
-from bloxlink_lib import redis
+from bloxlink_lib.database.redis import redis
 from bloxlink_lib.models.base.serializable import GuildSerializable, MemberSerializable
-from .guilds import *
-from .users import *
+from bloxlink_lib.models.schemas.guilds import *
+from bloxlink_lib.models.schemas.users import *
 
 
 async def fetch_item[T](domain: str, constructor: Type[T], item_id: str, *aspects) -> T:
@@ -18,24 +18,33 @@ async def fetch_item[T](domain: str, constructor: Type[T], item_id: str, *aspect
         item = await redis.hgetall(f"{domain}:{item_id}")
 
     if not item:
-        item = await mongo.bloxlink[domain].find_one({"_id": item_id}, {x: True for x in aspects}) or {
-            "_id": item_id
-        }
+        item = await mongo.bloxlink[domain].find_one(
+            {"_id": item_id}, {x: True for x in aspects}
+        ) or {"_id": item_id}
 
         if item and not isinstance(item, (list, dict)):
             if aspects:
-                items = {x: item[x] for x in aspects if item.get(
-                    x) and not isinstance(item[x], dict)}
+                items = {
+                    x: item[x]
+                    for x in aspects
+                    if item.get(x) and not isinstance(item[x], dict)
+                }
 
                 if items:
                     async with redis.pipeline() as pipeline:
                         await pipeline.hmset(f"{domain}:{item_id}", items)
-                        await pipeline.expire(f"{domain}:{item_id}", int(datetime.timedelta(hours=1).total_seconds()))
+                        await pipeline.expire(
+                            f"{domain}:{item_id}",
+                            int(datetime.timedelta(hours=1).total_seconds()),
+                        )
                         await pipeline.execute()
             else:
                 async with redis.pipeline() as pipeline:
                     await pipeline.hmset(f"{domain}:{item_id}", item)
-                    await pipeline.expire(f"{domain}:{item_id}", int(datetime.timedelta(hours=1).total_seconds()))
+                    await pipeline.expire(
+                        f"{domain}:{item_id}",
+                        int(datetime.timedelta(hours=1).total_seconds()),
+                    )
                     await pipeline.execute()
 
     if item.get("_id"):
@@ -81,7 +90,9 @@ async def update_item(domain: str, item_id: str, **aspects) -> None:
     if redis_set_aspects:
         async with redis.pipeline() as pipeline:
             await pipeline.hset(f"{domain}:{item_id}", mapping=redis_set_aspects)
-            await pipeline.expire(f"{domain}:{item_id}", int(datetime.timedelta(hours=1).total_seconds()))
+            await pipeline.expire(
+                f"{domain}:{item_id}", int(datetime.timedelta(hours=1).total_seconds())
+            )
             await pipeline.execute()
 
     if redis_unset_aspects:
@@ -93,7 +104,9 @@ async def update_item(domain: str, item_id: str, **aspects) -> None:
     )
 
 
-async def fetch_user_data(user: str | int | dict | MemberSerializable, *aspects) -> schemas.UserData:
+async def fetch_user_data(
+    user: str | int | dict | MemberSerializable, *aspects
+) -> schemas.UserData:
     """
     Fetch a full user from local cache, then redis, then database.
     Will populate caches for later access
@@ -109,7 +122,9 @@ async def fetch_user_data(user: str | int | dict | MemberSerializable, *aspects)
     return await fetch_item("users", schemas.UserData, user_id, *aspects)
 
 
-async def fetch_guild_data(guild: str | int | dict | GuildSerializable, *aspects) -> GuildData:
+async def fetch_guild_data(
+    guild: str | int | dict | GuildSerializable, *aspects
+) -> GuildData:
     """
     Fetch a full guild from local cache, then redis, then database.
     Will populate caches for later access
@@ -125,7 +140,9 @@ async def fetch_guild_data(guild: str | int | dict | GuildSerializable, *aspects
     return await fetch_item("guilds", GuildData, guild_id, *aspects)
 
 
-async def update_user_data(user: str | int | dict | MemberSerializable, **aspects) -> None:
+async def update_user_data(
+    user: str | int | dict | MemberSerializable, **aspects
+) -> None:
     """
     Update a user's aspects in local cache, redis, and database.
     """
@@ -140,7 +157,9 @@ async def update_user_data(user: str | int | dict | MemberSerializable, **aspect
     return await update_item("users", user_id, **aspects)
 
 
-async def update_guild_data(guild: str | int | dict | GuildSerializable, **aspects) -> None:
+async def update_guild_data(
+    guild: str | int | dict | GuildSerializable, **aspects
+) -> None:
     """
     Update a guild's aspects in local cache, redis, and database.
     """
