+++ bloxlink_lib/models/schemas/guilds/__init__.py
@@ -1,9 +1,14 @@
 from typing import Mapping, Self, Type, Literal, Annotated
 from pydantic import Field, field_validator
-from .base import PydanticList, BaseModel, PydanticDict, MemberSerializable
-from ..validators import is_positive_number_as_str
-from .migrators import migrate_restrictions, migrate_delete_commands
-import bloxlink_lib.models.binds as binds_module
+from bloxlink_lib.models.base import (
+    PydanticList,
+    BaseModel,
+    PydanticDict,
+    MemberSerializable,
+)
+from bloxlink_lib.validators import is_positive_number_as_str
+from bloxlink_lib.models.migrators import migrate_restrictions, migrate_delete_commands
+from bloxlink_lib.models.binds import GuildBind
 
 
 class UserInfoFieldMapping(BaseModel):
@@ -70,7 +75,7 @@ class GuildData(BaseModel):
 
     id: Annotated[int, Field(alias="_id")]
 
-    binds: Annotated[list[binds_module.GuildBind], Field(default_factory=list)]
+    binds: Annotated[list[GuildBind], Field(default_factory=list)]
 
     verifiedRoleEnabled: bool = True
     verifiedRoleName: str | None = "Verified"  # deprecated
@@ -122,11 +127,11 @@ class GuildData(BaseModel):
     # field converters
     @field_validator("binds", mode="before")
     @classmethod
-    def transform_binds(cls: Type[Self], binds: list) -> list[binds_module.GuildBind]:
-        if all(isinstance(b, binds_module.GuildBind) for b in binds):
+    def transform_binds(cls: Type[Self], binds: list) -> list[GuildBind]:
+        if all(isinstance(b, GuildBind) for b in binds):
             return binds
 
-        return [binds_module.GuildBind(**b) for b in binds]
+        return [GuildBind(**b) for b in binds]
 
     @field_validator("deleteCommands", mode="before")
     @classmethod
@@ -145,7 +150,7 @@ class GuildData(BaseModel):
     def model_post_init(self, __context):
         # merge verified roles into binds
         if self.verifiedRole:
-            verified_role_bind = binds_module.GuildBind(
+            verified_role_bind = GuildBind(
                 criteria={"type": "verified"}, roles=[self.verifiedRole]
             )
 
@@ -153,7 +158,7 @@ class GuildData(BaseModel):
                 self.binds.append(verified_role_bind)
 
         if self.unverifiedRole:
-            unverified_role_bind = binds_module.GuildBind(
+            unverified_role_bind = GuildBind(
                 criteria={"type": "unverified"}, roles=[self.unverifiedRole]
             )
 
@@ -171,4 +176,4 @@ class GuildData(BaseModel):
 
 
 # RoleSerializable is not defined when the schema is first built, so we need to re-build it. TODO: make better
-binds_module.GuildBind.model_rebuild()
+GuildBind.model_rebuild()
